Move AspNetCoreTestFixture to the Datadog.Trace.TestHelpers.AutoInstrumentation library
